Fixed find_packages 'bug'.
The default package setup seems to be looking for subpackages that include the main package name, which means that if the subpackage directories dont include the main package name then they will not be included. The fix is just to remove this inclusion constrainy by using find_packages() instead of find_packages(include='...').